+++ apps/api/gql/generated-schema-ast.ts
@@ -194,9 +194,11 @@ export const schema = gql`
         DIRECTIONAL_FEE
         EXIT_FEE
         FEE_TAKING
+        LBP
         LOTTERY
         MEV_TAX
         NFTLIQUIDITY_POSITION
+        RECLAMM
         STABLE_SURGE
         UNKNOWN
         VEBAL_DISCOUNT
@@ -445,6 +447,11 @@ export const schema = gql`
         dSq: String
 
         """
+        ReClamm: Represents how fast the pool can move the virtual balances per day
+        """
+        dailyPriceShiftBase: BigDecimal
+
+        """
         The decimals of the BPT, usually 18
         """
         decimals: Int!
@@ -538,11 +545,6 @@ export const schema = gql`
         ReClamm: The timestamp when the update begins
         """
         priceRatioUpdateStartTime: Int
-
-        """
-        ReClamm: Represents how fast the pool can move the virtual balances per day
-        """
-        priceShiftRatePerSecond: BigDecimal
 
         """
         The protocol version on which the pool is deployed, 1, 2 or 3
@@ -1948,6 +1950,11 @@ export const schema = gql`
         The current fourth root price ratio, an interpolation of the price ratio state
         """
         currentFourthRootPriceRatio: BigDecimal!
+
+        """
+        Represents how fast the pool can move the virtual balances per day
+        """
+        dailyPriceShiftBase: BigDecimal!
         decimals: Int!
         displayTokens: [GqlPoolTokenDisplay!]! @deprecated(reason: "Use poolTokens instead")
         dynamicData: GqlPoolDynamicData!
@@ -2002,11 +2009,6 @@ export const schema = gql`
         The timestamp when the update begins
         """
         priceRatioUpdateStartTime: Int!
-
-        """
-        Represents how fast the pool can move the virtual balances per day
-        """
-        priceShiftRatePerSecond: BigDecimal!
         protocolVersion: Int!
         staking: GqlPoolStaking
 
@@ -3900,12 +3902,22 @@ export const schema = gql`
         Pool ID
         """
         id: ID!
+
+        """
+        Returns all pool tokens, including BPTs and nested pools if there are any. Only one nested level deep.
+        """
+        poolTokens: [GqlPoolTokenDetail!]!
         protocolVersion: Int!
 
         """
         The symbol of the pool.
         """
         symbol: String!
+
+        """
+        List of tags assigned by the team based on external factors
+        """
+        tags: [String]
 
         """
         The tokens inside the pool.