+++ modules/pool/lib/apr-data-sources/morpho-api-client.ts
@@ -1,4 +1,5 @@
 import request, { gql } from 'graphql-request';
+import { applyOnchainDataUpdateCowAmm } from '../../../sources/enrichers/apply-onchain-data';
 
 const url = 'https://blue-api.morpho.org/graphql';
 const query = gql`
@@ -6,44 +7,50 @@ const query = gql`
         vaults(first: 1000, where: { netApy_gte: 0.00001 }) {
             items {
                 address
+                asset {
+                    address
+                    yield {
+                        apr
+                    }
+                }
                 chain {
                     network
                 }
                 state {
-                    apy
                     fee
-                    netApy
-                    rewards {
-                        supplyApr
-                        asset {
-                            name
-                            symbol
-                            address
-                        }
-                    }
+                    dailyApy
+                    dailyNetApy
                 }
             }
         }
     }
 `;
 
+/*
+Morpho APIs results are as follows:
+- dailyApy: Vault APY excluding rewards, before deducting the performance fee. Also NOT including the net APY of the underlying asset.
+- dailyNetApy: Vault APY including rewards and underlying yield, after deducting the performance fee.
+
+
+We only want to get the APY for rewards as we account for underlying yield separately inside the YB APR service. 
+We therefore deduct the fee from the apy and subtract the asset yield apr from it.
+*/
+
 type Vault = {
     address: string;
     chain: {
         network: string;
     };
+    asset: {
+        address: string;
+        yield?: {
+            apr: number;
+        };
+    };
     state: {
-        apy: number;
         fee: number;
-        netApy: number;
-        rewards: {
-            supplyApr: number;
-            asset: {
-                symbol: string;
-                name: string;
-                address: string;
-            };
-        }[];
+        dailyApy: number;
+        dailyNetApy: number;
     };
 };
 
@@ -59,62 +66,26 @@ const mapMorphoNetworkToChain = {
 };
 
 export const morphoApiClient = {
-    rewardTokens: async () => {
+    morphoApr: async () => {
         const {
             vaults: { items },
         } = await request<BlueApiResponse>(url, query);
 
-        // Map reward tokens to vault addresses
+        // Map apy to vault addresses
         return Object.fromEntries(
             items.map((vault: Vault) => [
                 vault.address.toLowerCase(),
-                vault.state.rewards.map((reward) => ({
-                    name: reward.asset.name,
-                    symbol: reward.asset.symbol,
-                    address: reward.asset.address.toLowerCase(),
-                    apr: reward.supplyApr,
+                {
+                    dailyApy: vault.state.dailyApy,
+                    dailyNetApy: vault.state.dailyNetApy,
+                    fee: vault.state.fee,
+                    rewardApy:
+                        vault.state.dailyNetApy -
+                        vault.state.dailyApy * (1 - vault.state.fee) -
+                        (vault.asset.yield?.apr || 0),
                     chain: mapMorphoNetworkToChain[vault.chain.network as keyof typeof mapMorphoNetworkToChain],
-                })),
+                },
             ]),
         );
-    },
-    morphoApr: async () => {
-        const query = gql`
-            {
-                mainnet: vault(id: "560b57bd-0e46-425f-9549-f3e38be0e1e6") {
-                    state {
-                        netApyWithoutRewards
-                        netApy
-                    }
-                }
-                base: vault(id: "5ebe85c0-0049-47bd-b6ab-ec913189191a") {
-                    state {
-                        netApyWithoutRewards
-                        netApy
-                    }
-                }
-            }
-        `;
-
-        type Response = {
-            mainnet: {
-                state: {
-                    netApyWithoutRewards: number;
-                    netApy: number;
-                };
-            };
-            base: {
-                state: {
-                    netApyWithoutRewards: number;
-                    netApy: number;
-                };
-            };
-        };
-
-        const r = await request<Response>(url, query);
-        const mainnetApr = r?.mainnet?.state?.netApy - r?.mainnet?.state?.netApyWithoutRewards;
-        const baseApr = r?.base?.state?.netApy - r?.base?.state?.netApyWithoutRewards;
-
-        return { MAINNET: mainnetApr, BASE: baseApr };
     },
 };